- More code analysis warnings, mostly looking at using the correct types (as usual). - Fixed a possible buffer overrun caused by passing the buffer length incorrectly to the EM_GETLINE message (fortunately, log lines would never have been long enough for this to have been an issue before). - Fixed "strip axis" command, checking for "id" twice.

+++ 0.56.b/common/CExpression.cpp
@@ -411,10 +411,10 @@ try_dec:
 		INTRINSIC_TYPE iIntrinsic = (INTRINSIC_TYPE) FindTableHeadSorted( pszArgs, sm_IntrinsicFunctions, COUNTOF(sm_IntrinsicFunctions)-1 );
 		if ( iIntrinsic >= 0 )
 		{
-			int iLen = strlen(sm_IntrinsicFunctions[iIntrinsic]);
+			size_t iLen = strlen(sm_IntrinsicFunctions[iIntrinsic]);
 			if ( pszArgs[iLen] == '(' )
 			{
-				pszArgs += (iLen+1);
+				pszArgs += (iLen + 1);
 				TCHAR * pszArgsNext;
 				Str_Parse( const_cast<TCHAR*>(pszArgs), &(pszArgsNext), ")" );
